requirements.txt
Flask==2.1.0
google-cloud-datastore==2.4.0
requests==2.27.1
google-auth==2.6.6